Handle column types and nulls in CSV import
To handle column types (on which SQLite is famously loose anyway, but
this gets us the right sorting), we need to create the table first, then
skip the header row with `--skip 1`.

To deal with empty strings instead of nulls, we need to issue manual
UPDATE statements to fix the columns we care about - typically nullable
counting columns, as we don't want empty strings at the top when sorting
by those columns descending.