Allow projecting queries onto format and gender
By default (or if all boxes are unchecked), select all genders and / or
formats. Create a table for each format, gender pair with the relevant
results, and convenient `innings`, `bowling_innings`, and `team_innings`
tables. It's still possible to use `women_test_batting_innings` and
friends directly, and then the headers will be irrelevant.